fix PIM bug

+++ buhockeystatssite/flask_app.py
@@ -127,8 +127,8 @@ def players():
                 if formData['position'] == 'skater':
                     dfStat = burb.dfSeasSkateWomens
                     pens = dfStat['pens'].str.split('/', expand=True)
-                    dfStat.loc[:, 'pen'] = pens[0].replace('—', np.nan)
-                    dfStat.loc[:, 'pim'] = pens[1].replace('—', np.nan)
+                    dfStat.loc[:, 'pen'] = pens[0].replace('—', np.nan).astype('Int64')
+                    dfStat.loc[:, 'pim'] = pens[1].replace('—', np.nan).astype('Int64')
                 elif formData['position'] == 'goalie':
                     dfStat = burb.dfSeasGoalieWomens
                     rec = dfStat.record.str.split('-', expand=True)
